+++ src/tool/cpp/basewinrt/base.h
@@ -4,7 +4,6 @@
 #include "base_macros.h"
 #include "base_types.h"
 #include "base_extern.h"
-#include "base_forward.h"
 #include "base_meta.h"
 #include "base_identity.h"
 #include "base_handle.h"
